uicommands: move some to ui/operation
