+++ src/code/uiCommands.ts
@@ -1,33 +1,17 @@
-import { WasabeeOp } from "./model";
+import type { WasabeeOp } from "./model";
 import { WasabeePortal } from "./model";
 import { WasabeeBlocker } from "./model";
 import { WasabeeMarker } from "./model";
 import ConfirmDialog from "./dialogs/confirmDialog";
-import MergeDialog from "./dialogs/mergeDialog";
 import { WasabeeMe } from "./model";
 import wX from "./wX";
-import { opPromise, GetWasabeeServer, locationPromise } from "./server";
-import AuthDialog from "./dialogs/authDialog";
-import {
-  getSelectedOperation,
-  makeSelectedOperation,
-  opsList,
-  removeOperation,
-  changeOpIfNeeded,
-  duplicateOperation,
-} from "./selectedOp";
+import { locationPromise } from "./server";
 
 import * as PortalUI from "./ui/portal";
-import {
-  displayError,
-  displayInfo,
-  displayWarning,
-  ServerError,
-} from "./error";
+import { displayInfo } from "./error";
 import { deleteDatabase } from "./db";
 import { constants } from "./static";
 import type { Wasabee } from "./init";
-import { getMe } from "./model/cache";
 
 export function sendLocation() {
   if (!WasabeeMe.isLoggedIn()) return;
@@ -138,157 +122,6 @@ export async function blockerAutomark(operation: WasabeeOp, first = true) {
   await blockerAutomark(operation, false);
 
   if (first) operation.endBatchMode();
-}
-
-export function zoomToOperation(operation: WasabeeOp) {
-  if (!operation) return;
-  const mbr = operation.mbr;
-  if (mbr && mbr.isValid()) {
-    window.map.fitBounds(mbr);
-  }
-}
-
-export async function updateLocalOp(local, remote) {
-  const so = getSelectedOperation();
-  if (!local) {
-    await remote.store();
-    return false;
-  }
-
-  // if selected op, use current selected op object
-  const op = local.ID != so.ID ? local : so;
-
-  if (op.lasteditid == remote.lasteditid) {
-    // nothing to do except .server update
-    if (op.server !== remote.server) {
-      op.server = remote.server;
-      await op.store();
-      return op.ID === so.ID;
-    }
-    return false;
-  }
-
-  // no changes
-  if (!op.checkChanges()) {
-    await remote.store();
-    // if selected op, reload from the new op
-    return remote.ID === so.ID;
-  }
-
-  // partial update on fields the server is always right
-  op.teamlist = remote.teamlist;
-  op.mergeZones(remote);
-  op.remoteChanged = true;
-  await op.store();
-
-  // In case of selected op, suggest merge to the user
-  if (so === op) {
-    const con = new MergeDialog({
-      opOwn: so,
-      opRemote: remote,
-    });
-    con.enable();
-  }
-
-  return false;
-}
-
-export async function fullSync() {
-  const so = getSelectedOperation();
-  const server = GetWasabeeServer();
-
-  try {
-    let reloadOpID = null;
-    const me = await getMe(true);
-    const opsID = new Set(me.Ops.map((o) => o.ID));
-
-    // delete operations absent from server unless the owner
-    const ol = await opsList();
-    const serverOps = [];
-    for (const opID of ol) {
-      const op = await WasabeeOp.load(opID);
-      if (op && op.server === server && !opsID.has(op.ID)) serverOps.push(op);
-    }
-    for (const op of serverOps) {
-      // if owned, duplicate the OP
-      if (op.isOwnedOp()) {
-        const newop = await duplicateOperation(op.ID);
-        newop.name = op.name;
-        await newop.store();
-        // if selected op, we reload the local duplicate
-        if (op.ID === so.ID) reloadOpID = newop.ID;
-      }
-      // skip hook (not needed)
-      await WasabeeOp.delete(op.ID);
-    }
-    if (serverOps.length > 0)
-      console.log(
-        "remove",
-        serverOps.map((op) => op.ID)
-      );
-
-    const promises: Promise<WasabeeOp>[] = [];
-    for (const opID of opsID) {
-      promises.push(opPromise(opID));
-    }
-    const ops = (await Promise.allSettled(promises))
-      .filter((p) => p.status === "fulfilled")
-      .map((p: PromiseFulfilledResult<WasabeeOp>) => p.value);
-    for (const newop of ops) {
-      const localOp = await WasabeeOp.load(newop.ID);
-      const reloadSO = await updateLocalOp(localOp, newop);
-      if (reloadSO) reloadOpID = so.ID;
-    }
-
-    // replace current op by the server version if any
-    if (reloadOpID) await makeSelectedOperation(reloadOpID);
-    // change op if the current does not exist anymore
-    else {
-      const op = await changeOpIfNeeded();
-      if (op !== so) zoomToOperation(op);
-    }
-
-    window.map.fire("wasabee:teams"); // if any team dialogs are open
-
-    displayInfo(wX("SYNC DONE"));
-  } catch (e) {
-    console.error(e);
-    if (e instanceof ServerError) displayError(e);
-    if (WasabeeMe.isLoggedIn()) displayWarning(wX("NOT_LOADED"));
-    else new AuthDialog().enable();
-  }
-  // update UI to reflect new ops list
-  window.map.fire("wasabee:fullsync");
-}
-
-export async function syncOp(opID: OpID) {
-  const localOp = await WasabeeOp.load(opID);
-  const remoteOp = await opPromise(opID);
-  if (remoteOp.lasteditid != localOp.lasteditid) {
-    if (!localOp.localchanged) {
-      await remoteOp.store();
-    } else {
-      const con = new MergeDialog({
-        opOwn: localOp,
-        opRemote: remoteOp,
-      });
-      con.enable();
-    }
-  }
-}
-
-export function deleteLocalOp(opname: string, opid: OpID) {
-  const con = new ConfirmDialog({
-    title: wX("REM_LOC_CP", { opName: opname }),
-    label: wX("YESNO_DEL", { opName: opname }),
-    type: "operation",
-    callback: async () => {
-      await removeOperation(opid);
-      const newop = await changeOpIfNeeded(); // fires ui events
-      zoomToOperation(newop);
-    },
-  });
-  con.enable();
 }
 
 export function clearAllData() {
